Isolate server and client components for login
Also ensure that logged in users won't see the login screen by redirecting to /chat when in a session
